Add tests for initHash, endHash, sendDataNoDisplay
